style: move stats logic into separate file
