+++ src/utils/zon/index.ts
@@ -1,3 +1,4 @@
 export * from "./types.ts";
 export * from "./path.ts";
+export * from "./stats.ts";
 export * from "./zon.ts";
